// src/encapsulation/metadata-injection-keys.enum.ts
export enum MetadataInjectionKeys {
    FILE_LOCATION = 'file:location',
    PARAM_TYPES = 'design:paramtypes',
    DI_TOKEN = 'di:token',
    DI_INJECT_PARAM_TOKENS = 'di:inject_param_tokens',
    DI_INJECT_PARAM_OPTIONS = 'di:inject_param_options',
    CONTROLLER_BASE_ROUTE = 'controller:base_route',
    CONTROLLER_ROUTES = 'controller:routes',
    ROUTE_PATH_PARAMS = 'route:path_params',
    ROUTE_QUERY_PARAMS = 'route:query_params',
    ROUTE_HEADER_PARAMS = 'route:header_params',
    ROUTE_BODY = 'route:body',
    MODEL_PROPERTIES = 'model:properties'
}